// src/app/models/prof.ts
export interface Prof {
  id:                number;
  name:              string;
  email:             string;
  email_verified_at: null;
  value:             string;
  created_at:        Date;
  updated_at:        Date;
  phone:             string;
  whatsapp:          string;
  address:           string;
  image:             string;
  isbanned:          number;
  role:              string;
  bio:               string;
  company_name:      string;
}